Adopt scenario documentation model
