--- .myteam/application-docs/.source.yml
@@ -1,5 +1,0 @@
-downloaded_at: '2026-04-21T01:33:16.841391+00:00'
-local_path: .myteam/application-docs
-ref: main
-repo: beanlab/rosters
-roster: project-management/application-docs
