--- .myteam/application-docs/collaboration.md
@@ -1,5 +1,0 @@
-## Collaboration
-
-ALWAYS review application-design changes with the user before
-proceeding further. These documents often benefit from discussion and
-iterative refinement.
