--- .myteam/application-docs/refactoring.md
@@ -1,17 +1,0 @@
-## Refactoring
-
-As applications grow, the design documents will need to be reorganized.
-
-Do not refactor these documents yourself.
-Delegate to the `application-docs/refactoring` role.
-
-Refactoring should be considered when:
-
-- a document grows beyond roughly 400 to 500 lines
-- a document mixes unrelated concerns
-- it is no longer obvious where a new interface change belongs
-- the same concept is described in multiple places
-- a document has become hard to scan, navigate, or maintain
-
-After refactoring is complete, re-read the relevant documents before
-continuing your task.
