--- .myteam/application-docs/review-checklist.md
@@ -1,11 +1,0 @@
-## Review Checklist
-
-Before finishing, verify the following:
-
-- The change describes an external interface, not an internal
-  implementation detail.
-- The correct document was updated.
-- Parent or index documents were updated if discoverability changed.
-- New information is placed with related material.
-- Open questions or uncertainties are called out clearly.
-- The resulting documents are coherent and easy to navigate.
