--- .myteam/application-docs/skill.md
@@ -1,29 +1,0 @@
----
-name: Application Design
-description: |
-  Instructions for how to create and maintain documents that describe
-  the intent and design of the application.
-  Load this skill if you need to create or modify anything in the
-  `application-design/` folder, or if you are asked to update the
-  application design.
----
-
-# Application Design
-
-We document the design and intent of the application in a document tree
-rooted at `application-design/`.
-
-If this folder does not yet exist, create it.
-
-The purpose of these documents is to describe the application's
-external behavior and operational contract, not its internal code
-structure.
-
-Read the following documents before making changes:
-
-- `intent.md`
-- `structure.md`
-- `change-workflow.md`
-- `refactoring.md`
-- `review-checklist.md`
-- `collaboration.md`
