--- .myteam/scenario/.source.yml
@@ -1,5 +1,0 @@
-downloaded_at: '2026-05-07T18:59:21.058888+00:00'
-local_path: .myteam/experimental/tyteam/scenario
-ref: main
-repo: beanlab/rosters
-roster: experimental/tyteam/scenario
