+++ docs/scenarios/admin/reports_status.md
@@ -1,0 +1,30 @@
+# Reports Status
+
+## Purpose
+
+Confirm bot health.
+
+---
+
+# Context
+
+The bot is running and admin command routing is available in the
+configured admin channel.
+
+---
+
+# Action
+
+An operator sends `!status` in the configured admin channel.
+
+---
+
+# Outcome
+
+The application replies with `I am alive. 🦆`.
+
+---
+
+# Related Scenarios
+
+- [Routes admin channel messages to commands](../routing/routes_admin_channel_messages_to_commands.md)
