+++ docs/scenarios/assignment_feedback/reports_missing_sections_as_unsatisfactory.md
@@ -1,0 +1,33 @@
+# Reports Missing Sections As Unsatisfactory
+
+## Purpose
+
+Grade absent report sections.
+
+---
+
+# Context
+
+An assignment feedback workflow is grading a markdown report against the
+rubric for the detected assignment.
+
+---
+
+# Action
+
+The uploaded report is missing sections required by the assignment
+rubric.
+
+---
+
+# Outcome
+
+The grading response includes explicit unsatisfactory rubric feedback
+for each missing required section.
+
+---
+
+# Related Scenarios
+
+- [Grades markdown report upload](grades_markdown_report_upload.md)
+- [Rejects invalid assignment feedback uploads](rejects_invalid_assignment_feedback_uploads.md)
