+++ docs/scenarios/startup/starts_bot_runtime.md
@@ -1,0 +1,41 @@
+# Starts Bot Runtime
+
+## Purpose
+
+Start the Discord bot.
+
+---
+
+# Context
+
+An operator has a valid runtime configuration available as a local JSON
+or YAML file, an `s3://` URI, or the `CONFIG_FILE_S3_PATH` environment
+variable. The configuration includes the required runtime sections and
+valid Discord, SQL, duck, tool, and reporting settings.
+
+---
+
+# Action
+
+The operator runs `python -m src.main` with optional `--config`,
+`--debug`, and `--log-path` arguments.
+
+---
+
+# Outcome
+
+The application loads and resolves the configuration, connects required
+runtime dependencies, initializes configured ducks and tools, starts the
+Discord bot loop, and sends `Duck online` to the configured admin
+channel when ready.
+
+If `--debug` is provided, runtime logging uses debug verbosity. If
+`--log-path` is provided, runtime logs are written to that path.
+
+---
+
+# Related Scenarios
+
+- [Rejects invalid runtime configuration](rejects_invalid_runtime_configuration.md)
+- [Loads runtime configuration](../configuration/loads_runtime_configuration.md)
+- [Resolves included configuration](../configuration/resolves_included_configuration.md)
